refactor(mutator set): rename active window to swbf suffix
"Active window" is an overloaded term, referring to two different
things, see #71 [1]. The rename clarifies that this struct refers
to the portion of the SWBF that is represented explicity.
Additionally, a generic type parameter is introduced which
determines the size of the suffix. (In future, this size may be
different from the window size.)

[1]: https://github.com/Neptune-Crypto/neptune-core/issues/71