refactor: make ExpectedUtox work with addresses
This is an experimental change that changes field `receiver_preimage`
to `receiver_privacy_digest` in `ExpectedUtxo`.

Previously `ExpectedUtxo` required possession of a `SpendingKey`.
With this change it only requires a `ReceiverAddress`.

The drawback is that it becomes necessary for recipient to scan all the
known wallet keys to find a match.

An advantage is that ExpectedUtxo could be sent in public and also
OnChain and OffChain utxo notification can use same/similar codepaths.

Many tests are still failing.  If we decide to pursue this, they would
need to be updated to provide receiver_privacy_digest instead of
receiver_preimage.