fix(`PrimitiveWitness`): Fix arbitrary primitive witness with timelocks
The issue was that the transaction kernel was being modified to set
the timestamp to a given value, but the native currency type scripts
witnesses depended on the old kernel mast hash, resulting in their
failure. The fix involves passing the timestamp as an argument to
an implementation of arbitrary for primitive witness.

Co-authored-by: Thorkil Værge <thor@neptune.cash>