removal_records_integrity: audit preloaded data
Add audit of preloaded data, without auditing at the end. This is done
this way because this check is unfortunately very expensive for the
removal records which contains a `ChunkDictionary` field which is a list
of variable-sized elements.

This check adds ~25% to the cost of this program, and the cost scales
linearly with the number of inputs.

A profile reveals the cost of this check, for both 2 and for inputs.

Cf. #190.