refactor: Dedicate function to hash-lock
The most used lock script type is the one where the owner proves
knowledge of the preimage of a digest that is hardcoded in the lock
script code. There was a convenience function to generate such a
`LockScript` object but this convenience function was living in
`[..]/wallet/address/common.rs`. This commit moves the convenience
function to `LockScript` and renames it to `hash_lock`.