peer_loop: Increase default tolerance to 1.000
Increasing this because we're adding more sanction/punishment options,
and we don't want to disconnect and ban too frivolously. We should add
ways to reward peers too, though.

Maybe if you get a new canonical block, or a new valid and favorable
block proposal from a, they should be rewarded?