debug(mine-loop): Filter out invalid or unconfirmable transactions
Helps to hunt down #249. This commit modifies the interface of
`get_transactions_for_block` such that
 - it is async;
 - it filters out invalid transactions (and logs an error message if
   any are encountered);
 - it filters out unconfirmable transactions (and logs a warning
   message if any are encountered).

This change addresses in part the error observed in #249, which comes
about as a result of populating the block proposal with bad
transactions from the mempool. However this is not a satisfying fix:
the bad transactions should not live in the mempool to begin with!
Hopefully, the extra debug information can point us in the right
direction the next time the error is observed.

Also:
 - Add `warn!` statements in mutator set logic regarding removal
   record validity.
 - Clarify some docstrings.