refactor(PotentialPeersState): Prefer candidates with new IPs
With this commit, the peer discovery logic now prefers peers with IPs
that the client does not already have a connection to. If there are no
other candidates, connections with repeated IPs will still be formed, as
the repeated IDs might constitute multiple machines behind the same
NAT'ed IPv4 address.

This closes #262.