feat: Add CLI command to broadcast tip request
The new command is designed to trigger a hang (#282) on demand.

Specifically, the CLI command is `neptune-cli broadcast-tip-request`
and upon invoking it the node will send a `BlockRequestByHash` to all
peers, where the associated data is its tip hash. All peers are
expected to reply with the given block, which will be validated before
being rejected because it is not new.