revert: Make mining single-threaded again
Reverts commit 330fda9fb66538e0a3251418b18315848fa8bb65 (which makes
guessing multi-threaded) and fixes resulting compilation errors.

This reversion is a response to Issue #282 where we observe that the
rayon parallelism in guessing seems to interfere with the rayon
parallelism in `triton_vm::verify`, causing the node to hang.