refactor(`mine_loop`): Inspect mempool after creating coinbase transaction
Deprecate `create_block_transaction_stateless`, which fixes the opposite
(and wrong!) order.

As a result, transactions that are announced while the coinbase
transaction is being produced, can be admitted into the block. Increases
time window for other clients to announce transactions to be included in
the next block. Note that other clients running the update-branch of
SingleProof have less work than the composers
who go through SingleProof to get a coinbase-tx. So with the same
hardware, other clients will finish their Update before the
composer fetches transactions from the mempool.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>