feat(primitive_witness): Add method for updating
Add method for updating PrimitiveWitness to be valid under a new (later)
mutator set. This code already existed in a test function. So I
extracted it from there and wrote a thorough test (I think) of it.

This functionality will be required when the client has recentely done a
PW->PC upgrade but a new block has come in, rendering the PC deprecated
and unusable since PC cannot be updated. The solution is to instead
update the PW and use it to construct a new PC.