fix: Don't fail PW-PC upgrade on race condition with block
Previously, a pritmitive witness to proof collection upgrade would fail
if a new block came in before this task was done. With this commit, the
initial primitive witness is updated to be valid under the new block,
and a new PW-PC upgrade is performed.

The proof collection itself cannot be updated to match a new mutator
set, only pritmitive witnesses and `SingleProof`s can.