feat(PrimitiveWitness): Allow updating wrt. MS update
Add function `PrimitiveWitness::update_with_new_ms_data` to allow a
valid primitive witness to be updated from one valid state to another,
given a mutator set update data structure. Most of these fields are
either unchanged or straight forward to update (if you understand the
mutator set data structure well). The most complicated part is updating
the witness data of the type scripts. The solution I went for is to not
update this witness data but instead redirive it, as was being done when
this data was originally constructed, for the old mutator set state.

This updating function does assume though that the lock script witness
data does not change when the mutator set is updated.