feat(rpc_server): Support external PoW guesser programs
Add two endpoints, one for exporting the PoW puzzle and one for
receiving a PoW puzzle solution. With these endpoints, external programs
like GPU hashers and mining pools can receive a PoW puzzle (challenge)
that they can work on.

In the current implementation, the node's wallet secret is used as the
guesser digest in the block header, meaning that the guesser-reward will
go to the wallet handled by the client.

With this change, `main_loop` now has two ways of handling a
self-guessed block, one from the client's own guesser, and one from the
RPC server. For this reason, the handling of a self-guessed block is
factored out to a separate function in main_loop, and this function gets
ia test.

All block proposals that are exported in this way are stored in RAM such
that a correct guess will be rewarded, even if the node has received a
new, more favorable, block proposal in the meantime.

`GlobalState` gets a new field for the current best block proposal and
for the exported block proposals. This new field is called
`mining_state`.